+++ app/Http/Controllers/FinancialYearController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Http\Requests\FinancialYearRequest;
 use App\Models\FinancialYear;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Validator;
 
 class FinancialYearController extends Controller
 {
@@ -15,12 +16,15 @@ class FinancialYearController extends Controller
      */
     public function index()
     {
+        return view('financialYear.index');
+    }
+
+    public function fetchFinancialYears(){
         $financialYears = FinancialYear::all();
-        return view('financialYear.index', [
+        return response()->json([
             'financialYears' => $financialYears,
         ]);
     }
-
     /**
      * Show the form for creating a new resource.
      *
@@ -28,9 +32,7 @@ class FinancialYearController extends Controller
      */
     public function create()
     {
-        return view('financialYear.create', [
-            'financialYear' => new FinancialYear(),
-        ]);
+
     }
 
     /**
@@ -39,10 +41,21 @@ class FinancialYearController extends Controller
      * @param  \Illuminate\Http\Request  $request
      * @return \Illuminate\Http\Response
      */
-    public function store(FinancialYearRequest $request)
+    public function store(Request $request)
     {
-        FinancialYear::create($request->all());
-        return redirect(route('financialYear.index'));
+        $validator = Validator::make($request->all(), [
+            'start_date' => 'required',
+            'end_date' => 'required',
+            'year_string' => 'required',
+            'status' => 'required',
+        ]);
+        if (!$validator->passes()){
+            return response()->json(['status' => 0, 'error' => $validator->errors()->toArray()]);
+        }
+        $financialYear = FinancialYear::create($request->all());
+        if ($financialYear){
+            return response()->json(['status' => 1, 'message' => 'Financial Year Added Successfully']);
+        }
     }
 
     /**
@@ -62,11 +75,23 @@ class FinancialYearController extends Controller
      * @param  \App\Models\FinancialYear  $financialYear
      * @return \Illuminate\Http\Response
      */
-    public function edit(FinancialYear $financialYear)
+    public function edit($financialYear)
     {
-        return view('financialYear.edit', [
-            'financialYear' => $financialYear,
-        ]);
+        $financialYear = FinancialYear::find($financialYear);
+        $status = $financialYear->getAttributes()['status'];
+        if ($financialYear){
+            return response()->json([
+                'status' => 200,
+                'financialYear' => $financialYear,
+                'status' => $status
+            ]);
+        }
+        else{
+            return response()->json([
+                'status' => 404,
+                'message' => 'Financial Year Not Found',
+            ]);
+        }
     }
 
     /**
@@ -76,10 +101,20 @@ class FinancialYearController extends Controller
      * @param  \App\Models\FinancialYear  $financialYear
      * @return \Illuminate\Http\Response
      */
-    public function update(FinancialYearRequest $request, FinancialYear $financialYear)
+    public function update(Request $request, $financialYear)
     {
+        $validator = Validator::make($request->all(), [
+            'start_date' => 'required',
+            'end_date' => 'required',
+            'year_string' => 'required',
+            'status' => 'required',
+        ]);
+        if (!$validator->passes()){
+            return response()->json(['status' => 0, 'error' => $validator->errors()->toArray()]);
+        }
+        $financialYear = FinancialYear::find($financialYear);
         $financialYear->update($request->all());
-        return redirect(route('financialYear.index'));
+        return response()->json(['status' => 1, 'message' => 'Financial Year Updated Successfully']);
     }
 
     /**
@@ -88,9 +123,19 @@ class FinancialYearController extends Controller
      * @param  \App\Models\FinancialYear  $financialYear
      * @return \Illuminate\Http\Response
      */
-    public function destroy(FinancialYear $financialYear)
+    public function destroy($financialYear)
     {
+        $financialYear = FinancialYear::find($financialYear);
+        if (!$financialYear){
+            return response()->json([
+                'status' => 0,
+                'message' => 'Financial Year not exist'
+            ]);
+        }
         $financialYear->delete();
-        return redirect(route('financialYear.index'));
+        return response()->json([
+            'status' => 1,
+            'message' => 'Financial Year Deleted Successfully'
+        ]);
     }
 }
